[rv32i] Enhancement: Refactoring Interrupt Interface
