+++ include/arch/core/rv32i/pic.h
@@ -33,19 +33,9 @@
  * @addtogroup rv32i-core-pic PIC
  * @ingroup rv32i-core
  *
- * @brief Programmble Interrupt Controller
+ * @brief Programmable Interrupt Controller
  */
 /**@{*/
-
-	/* Must come first. */
-	#define __NEED_CORE_REGS
-	#define __NEED_CORE_TYPES
-
-	#include <arch/core/rv32i/mcall.h>
-	#include <arch/core/rv32i/regs.h>
-	#include <arch/core/rv32i/types.h>
-	#include <nanvix/cc.h>
-	#include <errno.h>
 
 	/**
 	 * @name PIC Memory Map
@@ -57,171 +47,6 @@
 	#define PIC_MTIME_OFFSET    0x0000bff8 /**< Offset for mtime Register     */
 	/**@}*/
 
-	/**
-	 * @brief Number of interrupt requests (IRQs).
-	 */
-	#define RV32I_IRQ_NUM 12
-
-#ifndef _ASM_FILE_
-
-	/**
-	 * Table of Interrupt Requests (IRQs).
-	 */
-	extern const int irqs[RV32I_IRQ_NUM];
-
-	/**
-	 * @brief Unmasks an interrupt.
-	 *
-	 * @param irqnum Number of target interrupt request.
-	 *
-	 * @returns Upon successful completion, zero is returned. Upon
-	 * failure, a negative error code is returned instead.
-	 *
-	 * @author Pedro Henrique Penna
-	 */
-	static inline int rv32i_pic_unmask(int irqnum)
-	{
-		rv32i_word_t sie;
-
-		/* Invalid interrupt number. */
-		if ((irqnum < 0) || (irqnum >= RV32I_IRQ_NUM))
-			return (-EINVAL);
-
-		asm volatile (
-			"csrrs %0, sie, %1"
-			: "=r"(sie)
-			: "r"(irqs[irqnum])
-		);
-
-		return (0);
-	}
-
-	/**
-	 * @brief Masks an interrupt.
-	 *
-	 * @param irqnum Number of the target interrupt request.
-	 *
-	 * @returns Upon successful completion, zero is returned. Upon
-	 * failure, a negative error code is returned instead.
-	 *
-	 * @author Pedro Henrique Penna
-	 */
-	static inline int rv32i_pic_mask(int irqnum)
-	{
-		rv32i_word_t sie;
-
-		/* Invalid interrupt number. */
-		if ((irqnum < 0) || (irqnum >= RV32I_IRQ_NUM))
-			return (-EINVAL);
-
-		asm volatile (
-			"csrrc %0, sie, %1"
-			: "=r"(sie)
-			: "r"(irqs[irqnum])
-		);
-
-		return (0);
-	}
-
-	/**
-	 * @brief Acknowledges an interrupt.
-	 *
-	 * @param irqnum Number of the target interrupt request.
-	 *
-	 * @returns Upon successful completion, zero is returned. Upon
-	 * failure, a negative error code is returned instead.
-	 *
-	 * @author Pedro Henrique Penna
-	 */
-	static inline int rv32i_pic_ack(int irqnum)
-	{
-		rv32i_word_t sie;
-
-		/* Invalid interrupt number. */
-		if ((irqnum < 0) || (irqnum >= RV32I_IRQ_NUM))
-			return (-EINVAL);
-
-		asm volatile (
-			"csrrc %0, sip, %1"
-			: "=r"(sie)
-			: "r"(irqs[irqnum])
-		);
-
-		return (0);
-	}
-
-	/**
-	 * @brief Unmasks all interrupts.
-	 *
-	 * the rv32i_pic_unmask_all() unmasks all interrupt request lines
-	 * in the underlying rv32i core.
-	 *
-	 * @author Pedro Henrique Penna
-	 *
-	 * @todo FIXME we should do this atomically.
-	 */
-	static inline void rv32i_pic_unmask_all(void)
-	{
-		for (int i = 0; i < RV32I_IRQ_NUM; i++)
-			rv32i_pic_unmask(i);
-	}
-
-	/**
-	 * @brief Masks all interrupts.
-	 *
-	 * the rv32i_pic_mask_all() masks all interrupts request lines in
-	 * the underlying rv32i core.
-	 *
-	 * @author Pedro Henrique Penna
-	 *
-	 * @todo FIXME we should do this atomically.
-	 */
-	static inline void rv32i_pic_mask_all(void)
-	{
-		for (int i = 0; i < RV32I_IRQ_NUM; i++)
-			rv32i_pic_unmask(i);
-	}
-
-#endif
-
 /**@}*/
-
-/*============================================================================*
- * Exported Interface                                                         *
- *============================================================================*/
-
-/**
- * @cond rv32i
- */
-
-	/**
-	 * @name Exported Functions
-	 */
-	/**@{*/
-	#define __interrupt_mask   /**< interrupt_mask()   */
-	#define __interrupt_unmask /**< interrupt_unmask() */
-	/**@}*/
-
-#ifndef _ASM_FILE_
-
-	/**
-	 * @see rv32i_pic_mask().
-	 */
-	static inline int interrupt_mask(int irqnum)
-	{
-		return (rv32i_pic_mask(irqnum));
-	}
-
-	/**
-	 * @see rv32i_pic_unmask().
-	 */
-	static inline int interrupt_unmask(int irqnum)
-	{
-		return (rv32i_pic_unmask(irqnum));
-	}
-
-#endif
-
-/**@endcond*/
 
 #endif /* ARCH_CORE_RV32I_PIC_H_  */
